common: Add some more JSON primitives
We were missing the `short_channel_id_dir` helpers.
